Add authentication operation tests and auth reducer tests
This commit introduces the addition of auth operation tests and auth reducer tests. It also includes the setup of Axios mock adapter for testing an HTTP request and both the 'redux-thunk' and 'redux-mock-store' dependencies were added for testing redux thunk functions. The 'auth/login' action's error handling was also updated.